feat(ui): hide node footer if there is nothing to display
