+++ invokeai/frontend/web/src/features/nodes/components/CustomNodes.tsx
@@ -1,9 +1,9 @@
-import InvocationNode from './nodes/InvocationNode';
 import CurrentImageNode from './nodes/CurrentImageNode';
+import InvocationNodeWrapper from './nodes/InvocationNodeWrapper';
 import NotesNode from './nodes/NotesNode';
 
 export const nodeTypes = {
-  invocation: InvocationNode,
+  invocation: InvocationNodeWrapper,
   current_image: CurrentImageNode,
   notes: NotesNode,
 };
